requirements.txt
adversarial-vision-challenge==1.0.5
astroid==2.0.4
bson==0.5.6
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorama==0.4.0
crowdai-api==0.1.17
crowdai-repo2docker==0.5.0
cycler==0.10.0
dask==0.20.2
decorator==4.3.0
docker==3.5.1
docker-pycreds==0.3.0
escapism==1.0.0
Flask==1.0.2
foolbox==1.7.0
future==0.16.0
gitdb2==2.0.5
GitPython==2.1.11
idna==2.7
ipython-genutils==0.2.0
isort==4.3.4
ItsDangerous==1.0.0
Jinja2==2.10
joblib==0.13.2
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==3.0.2
mccabe==0.6.1
networkx==2.2
numpy==1.15.3
opencv-python==3.4.3.18
packaging==18.0
Pillow==5.3.0
pylint==2.1.1
pyparsing==2.2.2
pypiwin32==223
python-dateutil==2.7.3
python-gitlab==1.6.0
python-json-logger==0.1.9
PyWavelets==1.0.1
pywin32==224
PyYAML==3.13
pyzmq==17.1.2
redis==2.10.6
requests==2.20.0
ruamel.yaml==0.15.74
scikit-image==0.14.1
scikit-learn==0.21.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
smmap2==2.0.5
toolz==0.9.0
torch==0.4.1
torchfile==0.1.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.24
visdom==0.1.8.5
websocket-client==0.53.0
Werkzeug==0.14.1
wincertstore==0.2
wrapt==1.10.11
